Fixed incorrect handling of avro null default values.
If a default is specified for a union, the default must be an instance of the first type in the union. Thus the generated schema needs to check the default definition when generating the union schema.

Additionally added a Avro.NULL constant that should be used to indicate a avro null default.